Add module CompleteToDoList to manage completed tasks
